+++ cassandra/__init__.py
@@ -23,7 +23,7 @@ class NullHandler(logging.Handler):
 logging.getLogger('cassandra').addHandler(NullHandler())
 
 
-__version_info__ = (2, 0, 2, 'post')
+__version_info__ = (2, 1, 0, 'b1')
 __version__ = '.'.join(map(str, __version_info__))