+++ CHANGELOG.rst
@@ -38,6 +38,8 @@ Bug Fixes
 * Avoid UnicodeDecodeError when query string is unicode (PYTHON-76)
 * Correctly capture dclocal_read_repair_chance for tables and
   use it when generating CREATE TABLE statements (PYTHON-84)
+* Avoid race condition with AsyncoreConnection that may cause messages
+  to fail to be written until a new message is pushed
 
 2.0.2
 =====